refactor: point in-repo callers at frappe.website.api
Update the web form / website JS and templates, the unsubscribe and
data-deletion confirmation link generators (new emails now embed the
canonical paths; already-sent links resolve via the aliases) and test
imports to the canonical frappe.website.api paths and names.
