+++ frappe/public/js/frappe/utils/common.js
@@ -445,7 +445,7 @@ frappe.utils.new_auto_repeat_prompt = function (frm) {
 };
 
 frappe.utils.get_page_view_count = function (route) {
-	return frappe.call("frappe.website.doctype.web_page_view.web_page_view.get_page_view_count", {
+	return frappe.call("frappe.website.api.get_page_view_count", {
 		path: route,
 	});
 };
